06-openai-agents-sdk/pyproject.toml
[project]
name = "06-openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.0",
    "asyncio>=3.4.3",
    "openai>=1.82.0",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]
